Re-organize commands
